Moved Thread class to utils
